+++ test/strongbox/jfx_test.clj
@@ -1,6 +1,6 @@
 (ns strongbox.jfx-test
   (:require
-   ;;[clj-http.fake :refer [with-fake-routes-in-isolation]]
+   ;;[clj-http.fake :refer [with-global-fake-routes-in-isolation]]
    [clojure.test :refer [deftest testing is use-fixtures]]
    [strongbox.ui.jfx :as jfx]
    ;;[taoensso.timbre :as log :refer [debug info warn error spy]]