+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file. This change
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.5.0 - 2021-09-05
+
+### Added
+
+* better error handling when no network connection is available.
+* a warning when the update available is from a different addon host.
+    - for example, an addon is removed from curseforge and now receives updates on wowinterface exclusively.
+
+### Changed
+
+* addons that fail to match against the catalogue but have matched against a catalogue previously can now be updated.
+    - for example, addons in these cases are now updateable: 
+        - an addon doesn't get updates any more but still works and is eventually removed from the 'short' catalogue.
+        - you decide to go curseforge-only and have that one addon from wowinterface you can't live without (or vice-versa)
+
 ## 4.4.1
 
 ### Changed